refactor: api-client -> class-api-client
