--- .eslintingore
@@ -1,7 +1,0 @@
-lib
-debug
-test/code/
-lib/
-test/kubernetes
-test/argo-rollout
-CHANGELOG.md
