+++ src/code-templates/class-api-client/ApiClientArgument.ts
