+++ src/code-templates/class-api-client/ApiClientClass/ApiClientInterface.ts
