+++ src/code-templates/class-api-client/ApiClientClass/Class.ts
