+++ src/code-templates/class-api-client/ApiClientClass/MethodBody/PathParameter.ts
