+++ src/code-templates/class-api-client/ApiClientClass/MethodBody/__tests__/PathParameter-test.ts
