+++ src/code-templates/class-api-client/index.ts
